+++ lib/oanda_api/client/username_client.rb
@@ -1,5 +1,9 @@
 module OandaAPI
   module Client
+    #
+    # *DEPRECATED:* The Sandbox API endpoint is no longer supported by Oanda.
+    # @deprecated Please use {OandaAPI::Client::TokenClient} with a practice account instead.
+    #
     # Makes requests to the API.
     # Instances access the Oanda _sandbox_ environment.
     # Most client requests require a valid Oanda sandbox account username.
@@ -36,6 +40,7 @@ module OandaAPI
 
       # @param [String] username used for authentication.
       def initialize(username, options={})
+        warn Kernel.caller.first + " [DEPRECATION] `OandaAPI::Client::UsernameClient` is deprecated.  Please use `OandaAPI::Client::TokenClient` instead."
         super options
         @domain = :sandbox
         @username = username