+++ spec/oanda_api/examples/accounts_spec.rb
@@ -4,13 +4,6 @@ require 'support/client_helper'
 describe "OandaAPI::Resource::Account" do
   let(:client) { ClientHelper.client }
 
-  it "creates a new account" do
-    VCR.use_cassette("accounts.create") do
-      account = client.accounts.create
-      expect(account).to be_an OandaAPI::Resource::Account
-    end
-  end
-
   it "gets all accounts", :vcr do
     VCR.use_cassette("accounts.get") do
       accounts = client.accounts.get