+++ spec/oanda_api/examples/orders_spec.rb
@@ -3,10 +3,10 @@ require 'support/client_helper'
 
 describe "OandaAPI::Resource::Order" do
   let(:account) { ClientHelper.account }
-  let(:price)   { ClientHelper.instrument("EUR_USD").bid - 0.1 }
+  let(:price)  { ->(instrument) { ClientHelper.instrument(instrument).bid - 0.1 } }
 
   it "creates an order", :vcr do
-    VCR.use_cassette("account(id).order(options).create") do
+    VCR.use_cassette("order(options).create") do
       order = account.order(instrument: "EUR_USD",
                                   type: "market",
                                   side: "buy",
@@ -17,16 +17,16 @@ describe "OandaAPI::Resource::Order" do
   end
 
   it "gets all open orders", :vcr do
-    VCR.use_cassette("account(id).orders.get") do
-      ClientHelper.create_order(type: "limit", price: price)
+    VCR.use_cassette("orders.get") do
+      ClientHelper.create_order(type: "limit", instrument: "EUR_USD", price: price["EUR_USD"])
       orders = account.orders.get
       expect(orders.first).to be_an OandaAPI::Resource::Order
     end
   end
 
   it "gets a filtered list of open orders", :vcr do
-    VCR.use_cassette("account(id).orders(options).get") do
-      ClientHelper.create_order(instrument: "USD_JPY", type: "limit", price: price)
+    VCR.use_cassette("orders(options).get") do
+      ClientHelper.create_order(type: "limit", instrument: "USD_JPY", price: price["USD_JPY"])
 
       orders = account.orders(instrument: "USD_JPY").get
       expect(orders.first).to be_an OandaAPI::Resource::Order
@@ -34,8 +34,8 @@ describe "OandaAPI::Resource::Order" do
   end
 
   it "gets a specific open order", :vcr do
-    VCR.use_cassette("account(id).order(id).get") do
-      order_id = ClientHelper.create_order(type: "limit", price: price)
+    VCR.use_cassette("order(id).get") do
+      order_id = ClientHelper.create_order(type: "limit", instrument: "EUR_USD", price: price["EUR_USD"])
                              .order_opened
                              .id
       expect(account.order(order_id).get).to be_an OandaAPI::Resource::Order
@@ -43,8 +43,8 @@ describe "OandaAPI::Resource::Order" do
   end
 
   it "modifies an open order", :vcr do
-    VCR.use_cassette("account(id).order(options).update") do
-      order_id = ClientHelper.create_order(type: "limit", price: price)
+    VCR.use_cassette("order(options).update") do
+      order_id = ClientHelper.create_order(type: "limit", instrument: "EUR_USD", price: price["EUR_USD"])
                              .order_opened
                              .id
       updated_order = account.order(id: order_id, units: 9_000).update
@@ -53,8 +53,8 @@ describe "OandaAPI::Resource::Order" do
   end
 
   it "closes an open order", :vcr do
-    VCR.use_cassette("account(id).order(id).close") do
-      order_id = ClientHelper.create_order(type: "limit", price: price)
+    VCR.use_cassette("order(id).close") do
+      order_id = ClientHelper.create_order(type: "limit", instrument: "EUR_USD", price: price["EUR_USD"])
                              .order_opened
                              .id
       before_order_ids = account.orders.get.map(&:id)