+++ spec/oanda_api/examples/positions_spec.rb
@@ -5,7 +5,7 @@ describe "OandaAPI::Resource::Position" do
   let(:account) { ClientHelper.account }
 
   it "gets all open positions", :vcr do
-    VCR.use_cassette("account(id).positions.get") do
+    VCR.use_cassette("positions.get") do
       ClientHelper.create_trade
       positions = account.positions.get
       expect(positions.first).to be_an OandaAPI::Resource::Position
@@ -13,7 +13,7 @@ describe "OandaAPI::Resource::Position" do
   end
 
   it "gets the position for an instrument", :vcr do
-    VCR.use_cassette("account(id).positions(instrument).get") do
+    VCR.use_cassette("positions(instrument).get") do
       ClientHelper.create_trade(instrument: "USD_JPY")
       position = account.positions("USD_JPY").get
       expect(position).to be_an OandaAPI::Resource::Position
@@ -21,7 +21,7 @@ describe "OandaAPI::Resource::Position" do
   end
 
   it "closes an existing position", :vcr do
-    VCR.use_cassette("account(id).positions(instrument).close") do
+    VCR.use_cassette("positions(instrument).close") do
       ClientHelper.create_trade(instrument: "USD_JPY")
 
       # Verify we have an open position.