+++ spec/oanda_api/examples/trades_spec.rb
@@ -5,7 +5,7 @@ describe "OandaAPI::Resource::Trade" do
   let(:account) { ClientHelper.account }
 
   it "gets a list of trades", :vcr do
-    VCR.use_cassette("account(id).trades.get") do
+    VCR.use_cassette("trades.get") do
       ClientHelper.create_trade
       trades = account.trades.get
       expect(trades.first).to be_an OandaAPI::Resource::Trade
@@ -13,7 +13,7 @@ describe "OandaAPI::Resource::Trade" do
   end
 
   it "gets a filtered list of trades", :vcr do
-    VCR.use_cassette("account(id).trades(filter).get") do
+    VCR.use_cassette("trades(filter).get") do
       ClientHelper.create_trade(instrument: "USD_JPY")
       trades = account.trades(instrument: "USD_JPY").get
       expect(trades.first).to be_an OandaAPI::Resource::Trade
@@ -21,7 +21,7 @@ describe "OandaAPI::Resource::Trade" do
   end
 
   it "gets information on a specific trade", :vcr do
-    VCR.use_cassette("account(id).trade(id).get") do
+    VCR.use_cassette("trade(id).get") do
       trade_id = ClientHelper.create_trade.trade_opened.id
       trade = account.trade(trade_id).get
       expect(trade).to be_an OandaAPI::Resource::Trade
@@ -29,7 +29,7 @@ describe "OandaAPI::Resource::Trade" do
   end
 
   it "modifies an open trade " do
-    VCR.use_cassette("account(id).trade(options).modify") do
+    VCR.use_cassette("trade(options).modify") do
       order = ClientHelper.create_trade
       modified_order =
         account.trade(id: order.trade_opened.id, take_profit: order.price + 2.0)
@@ -39,7 +39,7 @@ describe "OandaAPI::Resource::Trade" do
   end
 
   it "closes an open trade" do
-    VCR.use_cassette("account(id).trade(id).close") do
+    VCR.use_cassette("trade(id).close") do
       ClientHelper.create_trade(instrument: "USD_JPY").trade_opened.id
 
       # Get the oldest open trade for USD_JPY, so we can close that trade.