+++ spec/oanda_api/examples/transactions_spec.rb
@@ -5,7 +5,7 @@ describe "OandaAPI::Transaction" do
   let(:account) { ClientHelper.account }
 
   it "gets transaction history" do
-    VCR.use_cassette("account(id).transactions(options).get") do
+    VCR.use_cassette("transactions(options).get") do
       ClientHelper.create_trade(instrument: "USD_JPY")
       transactions = account.transactions(instrument: "USD_JPY").get
       expect(transactions.first).to be_an OandaAPI::Resource::Transaction
@@ -13,7 +13,7 @@ describe "OandaAPI::Transaction" do
   end
 
   it "gets information for a specific transaction" do
-    VCR.use_cassette("account(id).transaction(id).get") do
+    VCR.use_cassette("transaction(id).get") do
       ClientHelper.create_trade(instrument: "USD_JPY")
       id = account.transactions(instrument: "USD_JPY").get.first.id
       transaction = account.transaction(id).get