+++ spec/support/vcr.rb
@@ -4,5 +4,6 @@ VCR.configure do |c|
   c.hook_into :webmock
   c.configure_rspec_metadata!
   c.default_cassette_options = { record: :new_episodes }
+  c.filter_sensitive_data('<OANDA_TESTING_API_TOKEN>') { ClientHelper.test_account_token }
   c.ignore_hosts 'codeclimate.com'
 end